Moved large_scale_density functions into their own dedicated sub-package
